move json corpus generator to ydb/library
