+++ ydb/library/json_index/ya.make
@@ -2,6 +2,7 @@ LIBRARY()
 
 SRCS(
     json_index.cpp
+    json_corpus.cpp
 )
 
 PEERDIR(
